+++ packages/plugin-sound-cloud/web/src/createSoundCloudPlugin.ts
@@ -8,7 +8,7 @@ const createSoundCloudPlugin: CreatePluginFunction<SoundCloudPluginEditorConfig>
   const type = SOUND_CLOUD_TYPE;
   const { helpers, t, [type]: settings = {}, isMobile, ...rest } = config;
   // eslint-disable-next-line no-console
-  console.error(
+  console.warn(
     // eslint-disable-next-line max-len
     `Sound-cloud plugin is deprecated and will be removed in the next version. Please consume sound-cloud plugin through video plugin as described in  https://ricos.js.org/docs/plugins_api/VideoPlugin#sound-cloud-plugin`
   );